-yarn build + check-all test -fix stopInput bug -gives better names to some params
