+++ .eslintrc.cjs
@@ -10,9 +10,11 @@ module.exports = {
   plugins: ["react-refresh", "react-hooks"],
   rules: {
     "react-refresh/only-export-components": "warn",
-    test: /.ts$/,
-    exclude: /node_modules/,
-    loader: "ts-loader",
+    "@typescript-eslint/no-explicit-any": ["off"],
+    "@typescript-eslint/ban-ts-comment": ["off"],
+    test: 0,
+    exclude: 0,
+    loader: 0,
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
   },
