+++ index.html
@@ -5,11 +5,11 @@
     <link rel="icon" type="image" href="src\assets\images\elma_cute.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Elma GPT</title>
+    <script src="src\live2d.min.js"></script>
+    <script src="src\live2dcubismcore.min.js"></script>
   </head>
   <body>
     <div id="root"></div>
-    <script src="src\live2d.min.js"></script>
-    <script src="src\live2dcubismcore.min.js"></script>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
