+++ package.json
@@ -52,7 +52,7 @@
     "husky": "^8.0.0",
     "lint-staged": "^13.2.3",
     "postcss": "^8.4.24",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.1",
     "tailwindcss": "^3.3.2",
     "ts-loader": "^9.4.4",
     "typescript": "^5.0.2",
