+++ src/Global/module.ts
@@ -1,0 +1,1 @@
+declare module "chinese-conv";
